Split annotations into their own files.
